server/requirements.txt
fastapi
uvicorn
python-multipart
langchain-google-genai
langchain-community
langchain
langchain-chroma
chromadb
pypdf
python-dotenv
langchain-mongodb
pymongo
motor
aiofiles
pydantic
sentence-transformers
pdfminer
unstructured
pymupdf
pydantic[email]
fastapi_mail
wikipedia
duckduckgo-search
langchain-tools
ddgs
python-jose[cryptography]
passlib[bcrypt]
argon2-cffi==23.1.0
